+++ lib/src/job/entity/job.dart
@@ -3,6 +3,7 @@
 // BSD-style license that can be found in the LICENSE file.
 
 // Project imports:
+import 'package:batch/src/job/context/execution_context.dart';
 import 'package:batch/src/job/entity/entity.dart';
 import 'package:batch/src/job/entity/step.dart';
 import 'package:batch/src/job/precondition.dart';
@@ -19,7 +20,19 @@ class Job extends Entity<Job> {
     required String name,
     this.schedule,
     Precondition? precondition,
-  }) : super(name: name, precondition: precondition);
+    Function(ExecutionContext context)? onStarted,
+    Function(ExecutionContext context)? onSucceeded,
+    Function(ExecutionContext context, dynamic error, StackTrace stackTrace)?
+        onFailed,
+    Function(ExecutionContext context)? onCompleted,
+  }) : super(
+          name: name,
+          precondition: precondition,
+          onStarted: onStarted,
+          onSucceeded: onSucceeded,
+          onFailed: onFailed,
+          onCompleted: onCompleted,
+        );
 
   /// The schedule
   final ScheduleParser? schedule;