+++ lib/src/job/entity/step.dart
@@ -3,6 +3,7 @@
 // BSD-style license that can be found in the LICENSE file.
 
 // Project imports:
+import 'package:batch/src/job/context/execution_context.dart';
 import 'package:batch/src/job/entity/entity.dart';
 import 'package:batch/src/job/entity/task.dart';
 import 'package:batch/src/job/precondition.dart';
@@ -13,7 +14,19 @@ class Step extends Entity<Step> {
   Step({
     required String name,
     Precondition? precondition,
-  }) : super(name: name, precondition: precondition);
+    Function(ExecutionContext context)? onStarted,
+    Function(ExecutionContext context)? onSucceeded,
+    Function(ExecutionContext context, dynamic error, StackTrace stackTrace)?
+        onFailed,
+    Function(ExecutionContext context)? onCompleted,
+  }) : super(
+          name: name,
+          precondition: precondition,
+          onStarted: onStarted,
+          onSucceeded: onSucceeded,
+          onFailed: onFailed,
+          onCompleted: onCompleted,
+        );
 
   /// The tasks
   final List<Task> tasks = [];