+++ lib/src/job/launcher/job_launcher.dart
@@ -6,6 +6,7 @@
 import 'package:batch/src/job/context/execution_context.dart';
 import 'package:batch/src/job/entity/job.dart';
 import 'package:batch/src/job/launcher/launcher.dart';
+import 'package:batch/src/job/launcher/step_launcher.dart';
 
 /// This class provides the feature to securely execute registered jobs.
 class JobLauncher extends Launcher<Job> {
@@ -19,5 +20,12 @@ class JobLauncher extends Launcher<Job> {
   final Job _job;
 
   @override
-  Future<void> run() async => await super.executeRecursively(entity: _job);
+  Future<void> run() async => await super.executeRecursively(
+        entity: _job,
+        execute: (job) async => await StepLauncher(
+          context: context,
+          steps: job.steps,
+          parentJobName: job.name,
+        ).run(),
+      );
 }