+++ lib/src/job/launcher/launcher.dart
@@ -7,10 +7,6 @@ import 'package:batch/src/job/branch/branch_status.dart';
 import 'package:batch/src/job/context/context_helper.dart';
 import 'package:batch/src/job/context/execution_context.dart';
 import 'package:batch/src/job/entity/entity.dart';
-import 'package:batch/src/job/entity/job.dart';
-import 'package:batch/src/job/entity/step.dart';
-import 'package:batch/src/job/launcher/step_launcher.dart';
-import 'package:batch/src/job/launcher/task_launcher.dart';
 import 'package:batch/src/log/logger_provider.dart';
 import 'package:batch/src/runner.dart';
 
@@ -21,44 +17,40 @@ abstract class Launcher<T extends Entity<T>> extends ContextHelper<T>
     required ExecutionContext context,
   }) : super(context: context);
 
-  Future<void> executeRecursively({required T entity}) async {
+  Future<void> executeRecursively({
+    required T entity,
+    required Function(dynamic entity) execute,
+  }) async {
     if (!entity.canLaunch()) {
       info('Skipped ${entity.name} because the precondition is not met.');
       return;
     }
 
-    await _execute(entity: entity);
+    super.startNewExecution(name: entity.name);
+
+    try {
+      entity.onStarted?.call(context);
+
+      await execute.call(entity);
+
+      entity.onSucceeded?.call(context);
+    } catch (error, stackTrace) {
+      entity.onFailed?.call(context, error, stackTrace);
+    } finally {
+      entity.onCompleted?.call(context);
+    }
+
+    super.finishExecution();
 
     if (entity.hasBranch) {
       for (final branch in entity.branches) {
         if (branch.on == super.branchStatus ||
             branch.on == BranchStatus.completed) {
-          await executeRecursively(entity: branch.to);
+          await executeRecursively(entity: branch.to, execute: execute);
         }
       }
     }
 
     super.resetBranchStatus();
-  }
-
-  Future<void> _execute({required dynamic entity}) async {
-    super.startNewExecution(name: entity.name);
-
-    if (entity is Job) {
-      await StepLauncher(
-        context: super.context,
-        steps: entity.steps,
-        parentJobName: entity.name,
-      ).run();
-    } else if (entity is Step) {
-      await TaskLauncher(
-        context: super.context,
-        tasks: entity.tasks,
-      ).run();
-    } else {
-      await entity.execute(super.context);
-    }
-
-    super.finishExecution();
   }
 }