+++ lib/src/job/launcher/step_launcher.dart
@@ -6,6 +6,7 @@
 import 'package:batch/src/job/context/execution_context.dart';
 import 'package:batch/src/job/entity/step.dart';
 import 'package:batch/src/job/launcher/launcher.dart';
+import 'package:batch/src/job/launcher/task_launcher.dart';
 
 class StepLauncher extends Launcher<Step> {
   /// Returns the new instance of [StepLauncher].
@@ -27,7 +28,13 @@ class StepLauncher extends Launcher<Step> {
     }
 
     for (final step in _steps) {
-      await super.executeRecursively(entity: step);
+      await super.executeRecursively(
+        entity: step,
+        execute: (step) async => await TaskLauncher(
+          context: context,
+          tasks: step.tasks,
+        ).run(),
+      );
     }
   }
 }